+++ app/teste/page.tsx
@@ -1,40 +1,12 @@
-"use client";
-
-import type { Photo } from "@/components/teste/PhotoGrid";
-import { PhotoGrid } from "@/components/teste/PhotoGrid";
 import dados from "./data.json";
+import ListPhoto from "./listPhoto";
 
 export default function Home() {
-    const handleOrderChange = async (newOrder: Photo[]) => {
-        try {
-            // Aqui você implementa a chamada à sua API para salvar a nova ordem
-            const response = await fetch("/api/photos/reorder", {
-                method: "POST",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ newOrder }),
-            });
-
-            if (response.ok) {
-                console.log("Ordem das fotos atualizada com sucesso");
-            }
-
-            if (!response.ok) {
-                throw new Error("Falha ao atualizar a ordem das fotos");
-            }
-        } catch (error) {
-            console.error("Erro ao salvar a nova ordem:", error);
-            // Aqui você pode implementar uma notificação de erro para o usuário
-        }
-    };
-
     return (
-        <div className="container mx-auto py-8">
+        <div className="container flex flex-col w-full h-screen bg-slate-300 mx-auto py-8 items-center">
             <h1 className="text-2xl font-bold mb-6">Relatório Fotográfico</h1>
-            <PhotoGrid
+            <ListPhoto
                 initialPhotos={dados.photos}
-                onOrderChange={handleOrderChange}
             />
         </div>
     );
